+++ src/agents/memory/index.ts
@@ -1,6 +1,6 @@
 import type { Message } from 'neuri/openai'
-import type { Action } from 'src/libs/mineflayer'
-import type { AgentConfig, MemoryAgent } from '../../libs/mineflayer/interfaces/agents'
+import type { Action } from '../../libs/mineflayer'
+import type { AgentConfig, MemoryAgent } from '../../libs/mineflayer/base-agent'
 import { useLogg } from '@guiiai/logg'
 import { Memory } from '../../libs/mineflayer/memory'
