+++ src/agents/planning/llm.ts
@@ -2,7 +2,7 @@ import type { Neuri } from 'neuri'
 import type { Action } from '../../libs/mineflayer/action'
 import { useLogg } from '@guiiai/logg'
 import { system, user } from 'neuri/openai'
-import { genPlanningAgentPrompt } from '../../prompts/agent'
+import { genPlanningAgentPrompt } from '../../utils/prompt'
 import { toRetriable } from '../../utils/reliability'
 
 const logger = useLogg('planning-llm').useGlobalConfig()
