+++ src/main.ts
@@ -1,5 +1,4 @@
 import process, { exit } from 'node:process'
-
 import { useLogg } from '@guiiai/logg'
 import { Client } from '@proj-airi/server-sdk'
 import MineflayerArmorManager from 'mineflayer-armor-manager'
@@ -9,11 +8,11 @@ import { pathfinder as MineflayerPathfinder } from 'mineflayer-pathfinder'
 import { plugin as MineflayerPVP } from 'mineflayer-pvp'
 import { plugin as MineflayerTool } from 'mineflayer-tool'
 
-import { initAgent } from './agents/action/openai'
+import { initAgent } from './agents/action/llm'
 import { initBot } from './composables/bot'
 import { botConfig, initEnv } from './composables/config'
-import { wrapPlugin } from './libs/mineflayer/plugin'
-import { LLMAgent } from './mineflayer/llm-agent'
+import { wrapPlugin } from './libs/mineflayer'
+import { LLMAgent } from './plugins/llm-agent'
 import { initLogger } from './utils/logger'
 
 const logger = useLogg('main').useGlobalConfig()
@@ -36,7 +35,7 @@ async function main() {
 
   const airiClient = new Client({ name: 'minecraft-bot', url: 'ws://localhost:6121/ws' })
 
-  // Dynamically load LLMAgent after bot is initialized
+  // Dynamically load LLMAgent after the bot is initialized
   const agent = await initAgent(bot)
   await bot.loadPlugin(LLMAgent({ agent, airiClient }))
