+++ src/plugins/pathfinder.ts
